Restructure navigation (#126)
