+++ web/src/i18n/locales/default.en-US.json
@@ -12,6 +12,7 @@
     "nav.processing": "Processing",
     "nav.analytics": "Analytics",
     "nav.fingerprints": "My Fingerprints",
+    "nav.dashboard": "Dashboard",
     "collection.title": "My Collection",
     "collection.dashboard.title": "Dashboard",
     "period.day": "Day",
